+++ src/common/components/card/CardViewer.js
@@ -25,7 +25,8 @@ export default class CardViewer extends Component {
         <CSSTransitionGroup
           transitionName="card-viewer-fade"
           transitionEnterTimeout={100}
-          transitionLeaveTimeout={100}>
+          transitionLeaveTimeout={100}
+        >
           {
             this.props.hoveredCard &&
               <Card
