+++ src/common/components/hexgrid/Hex.js
