+++ src/common/components/hexgrid/HexGridImages.js
