+++ src/common/components/hexgrid/HexUtils.js
