+++ src/common/components/hexgrid/Layout.js
