+++ src/common/components/hexgrid/Orientation.js
