+++ src/common/components/hexgrid/Path.js
