+++ src/common/components/hexgrid/Point.js
