+++ test/containers/GameArea.spec.js
@@ -8,8 +8,8 @@ import Board from '../../src/common/components/game/Board';
 import PlayerArea from '../../src/common/components/game/PlayerArea';
 import Status from '../../src/common/components/game/Status';
 import VictoryScreen from '../../src/common/components/game/VictoryScreen';
-import HexGrid from '../../src/common/components/react-hexgrid/HexGrid';
-import HexUtils from '../../src/common/components/react-hexgrid/HexUtils';
+import HexGrid from '../../src/common/components/hexgrid/HexGrid';
+import HexUtils from '../../src/common/components/hexgrid/HexUtils';
 import * as actions from '../../src/common/actions/game';
 import gameReducer from '../../src/common/reducers/game';
